# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "YourMusawo",
    "description": "Get Medical Treatment at the Comfort of your Home",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:4000",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/patients": {
      "post": {
        "tags": [
          "Client Account"
        ],
        "description": "Create new client/patient in system, required fields: name, phone, location, password, date of birth",
        "parameters": [
          {
            "name": "patients",
            "in": "body",
            "description": "Clients that we want to create an account",
            "schema": {
              "$ref": "#/definitions/patient"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Client/patient successfully created",
            "schema": {
              "$ref": "#/definitions/patient"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Client Account"
        ],
        "summary": "Get all clients in system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/patient"
            }
          }
        }
      }
    },
...
    "/patients/{userId}": {
      "get": {
        "tags": [
          "Client Account"
        ],
        "summary": "Get client with given ID",
        "responses": {
          "200": {
            "description": "User is found",
            "schema": {
              "$ref": "#/definitions/patient"
            }
          }
        }
      },
      "delete": {
        "summary": "Delete client with given ID",
        "tags": [
          "Client Account"
        ],
        "responses": {
          "200": {
            "description": "Client is deleted",
            "schema": {
              "$ref": "#/definitions/patient"
            }
          }
        }
      },
      "put": {
        "summary": "Update user with give ID",
        "tags": [
          "Client Account"
        ],
        "parameters": [
          {
            "name": "patients",
            "in": "body",
            "description": "Client with new values of properties: name. location, password",
            "schema": {
              "$ref": "#/definitions/patient"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "client is updated",
            "schema": {
              "$ref": "#/definitions/patient"
            }
          }
        }
      }
    },
    "/doctors": {
      "post": {
        "tags": [
          "Medical Personnel Account"
        ],
        "description": "Create new medical personnel in system, required fields: imageURL, imageId, firstName, lastName, nin, specialization, postAddress, dob, city, phone, password, priceRate",
        "parameters": [
          {
            "name": "doctors",
            "in": "body",
            "description": "Medical personnel that  want to create an account",
            "schema": {
              "$ref": "#/definitions/doctor"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Medical Personnel successfully created",
            "schema": {
              "$ref": "#/definitions/doctor"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Medical Personnel Account"
        ],
        "summary": "Get all medical personnel in system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/doctor"
            }
          }
        }
      }
    },
...
    "/doctors/{userId}": {
      "get": {
        "tags": [
          "Medical Personnel Account"
        ],
        "summary": "Get medical personnel with given ID",
        "responses": {
          "200": {
            "description": "User is found",
            "schema": {
              "$ref": "#/definitions/doctor"
            }
          }
        }
      },
      "delete": {
        "summary": "Delete medical personnel with given ID",
        "tags": [
          "Medical Personnel Account"
        ],
        "responses": {
          "200": {
            "description": "Medical Personnel is deleted",
            "schema": {
              "$ref": "#/definitions/doctor"
            }
          }
        }
      },
      "put": {
        "summary": "Update medical personnel with give ID",
        "tags": [
          "Medical Personnel Account"
        ],
        "parameters": [
          {
            "name": "doctor",
            "in": "body",
            "description": "Medical personnel with new values of properties: city. postAddress, password, priceRate, isAvailable, isActive",
            "schema": {
              "$ref": "#/definitions/doctor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "medical personnel is updated",
            "schema": {
              "$ref": "#/definitions/doctor"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "patient": {
      "required": [
        "name",
        "phone",
        "location",
        "password",
        "dob"
      ],
      "properties": {
        "name": {
          "type": "string",
          "uniqueItems": false
        },
        "phone": {
          "type": "string",
          "uniqueItems": true
        },
        "location": {
          "type": "string",
          "uniqueItems": false
        },
        "password": {
          "type": "string",
          "uniqueItems": false
        },
        "dob": {
          "type": "Date",
          "uniqueItems": false
        },
        "registeredOn": {
          "type": "boolean"
        }
      }
    },
    "doctor": {
      "required": [
        "imageURL",
        "imageId",
        "firstName",
        "lastName",
        "nin",
        "specialization",
        "postAddress",
        "dob",
        "city",
        "phone",
        "password",
        "priceRate"
      ],
      "properties": {
        "imageURL": {
          "type": "string",
          "uniqueItems": true
        },
        "imageId": {
          "type": "string",
          "uniqueItems": true
        },
        "firstName": {
          "type": "string",
          "uniqueItems": false
        },
        "lastName": {
          "type": "string",
          "uniqueItems": false
        },
        "nin": {
          "type": "string",
          "uniqueItems": true
        },
        "specialization": {
          "type": "string",
          "uniqueItems": false
        },
        "postAddress": {
          "type": "string",
          "uniqueItems": false
        },
        "city": {
          "type": "string",
          "uniqueItems": false
        },
        "priceRate": {
          "type": "string",
          "uniqueItems": false
        },
        "phone": {
          "type": "string",
          "uniqueItems": true
        },
        "location": {
          "type": "string",
          "uniqueItems": false
        },
        "password": {
          "type": "string",
          "uniqueItems": false
        },
        "dob": {
          "type": "Date",
          "uniqueItems": false
        },
        "isAvailable": {
          "type": "boolean"
        },
        "isActive": {
          "type": "boolean"
        },
        "registeredOn": {
          "type": "date",
          "default": "true"
        },
        "isAdmin": {
          "type": "boolean"
        },
        "everHired": {
          "type": "boolean"
        }
      }
    },
...
  }
}